+++ test/moonwall.config.json
@@ -8,7 +8,7 @@
             "name": "dev_tanssi",
             "timeout": 120000,
             "envVars": ["DEBUG_COLORS=1"],
-            "testFileDir": ["suites/common-all", "suites/common-xcm", "suites/common-tanssi", "suites/dev-tanssi"],
+            "testFileDir": ["suites/common-all", "suites/common-xcm", "suites/common-tanssi-parachain", "suites/common-tanssi", "suites/dev-tanssi"],
             "runScripts": [
                 "compile-wasm.ts compile -b ../target/release/tanssi-node -o wasm -c dancebox_dev -a='--mock-container-chain=2000 --mock-container-chain=2001'"
             ],
@@ -43,7 +43,7 @@
         {
             "name": "dev_tanssi_relay",
             "envVars": ["DEBUG_COLORS=1"],
-            "testFileDir": ["suites/dev-tanssi-relay", "suites/common-all"],
+            "testFileDir": ["suites/dev-tanssi-relay", "suites/common-all", "suites/common-tanssi"],
             "multiThreads": true,
             "reporters": ["basic"],
             "foundation": {
@@ -743,7 +743,7 @@
         {
             "name": "dev_flashbox",
             "envVars": ["DEBUG_COLORS=1"],
-            "testFileDir": ["suites/common-all",  "suites/common-all-parachain", "suites/common-tanssi"],
+            "testFileDir": ["suites/common-all",  "suites/common-all-parachain", "suites/common-tanssi-parachain", "suites/common-tanssi"],
             "runScripts": [
                 "compile-wasm.ts compile -b ../target/release/tanssi-node -o wasm -c flashbox_dev -a='--mock-container-chain=2000 --mock-container-chain=2001'"
             ],
