+++ test/suites/common-tanssi-parachain/registrar/test_registrar_register.ts
