+++ test/suites/common-tanssi-parachain/services-payment/test_service_payment_removes_tank_money_and_burns.ts
