+++ test/suites/common-tanssi/pallet-treasury/test_pallet_treasury.ts
@@ -2,7 +2,7 @@ import "@tanssi/api-augment";
 import { describeSuite, expect, beforeAll } from "@moonwall/cli";
 import { ApiPromise } from "@polkadot/api";
 import { KeyringPair } from "@moonwall/util";
-import { extractFeeAuthor } from "util/block";
+import { jumpBlocks } from "util/block";
 
 describeSuite({
     id: "CT0901",
@@ -14,9 +14,10 @@ describeSuite({
         let sudo_alice: KeyringPair;
         let user_charlie: KeyringPair;
         let user_dave: KeyringPair;
-        let user_bob: KeyringPair;
-        // From Pallet Id "tns/tsry" -> Account
-        const treasury_address = "5EYCAe5jXiVvytpxmBupXPCNE9Vduq7gPeTwy9xMgQtKWMnR";
+        let user_bob: KeyringPair;        
+        // From Pallet Id "py/trsry" -> Account if relay chain
+        // From Pallet Id "tns/tsry" -> Account if parachain
+        
 
         beforeAll(async () => {
             polkadotJs = context.polkadotJs();
@@ -24,38 +25,60 @@ describeSuite({
             user_charlie = context.keyring.charlie;
             user_dave = context.keyring.dave;
             user_bob = context.keyring.bob;
+            const runtimeName = polkadotJs.runtimeVersion.specName.toString();
+            const treasuryAddress = runtimeName.includes("light") ? "5EYCAe5ijiYfyeZ2JJCGq56LmPyNRAKzpG4QkoQkkQNB5e6Z" : "5EYCAe5jXiVvytpxmBupXPCNE9Vduq7gPeTwy9xMgQtKWMnR";
+            const signedTx = await polkadotJs.tx.balances.transferAllowDeath(treasuryAddress, 1_000_000_000_000).signAsync(sudo_alice);
+            // Fund treasury account
+            await context.createBlock([signedTx]);
         });
 
         it({
             id: "E01",
-            title: "20% of fees & tips go for treasury account",
+            title: "Non root can not spend from treasury (Local)",
             test: async function () {
-                // Gets the initial pot deposit value
-                const initial_pot = await polkadotJs.query.system.account(treasury_address);
-                const initial_free_pot = initial_pot.data.free.toBigInt();
+                expect((await polkadotJs.query.treasury.spendCount()).toNumber()).to.equal(0);
 
-                // Executes a tx adding an additional tip
-                const tx = polkadotJs.tx.balances.transferAllowDeath(user_charlie.address, 200_000);
-                const signedTx = await tx.signAsync(user_dave, { tip: 100_000 });
+                // Creates a proposal
+                const proposal_value = 1000000000n;
+                const tx = polkadotJs.tx.treasury.spendLocal(proposal_value, user_dave.address);
+                const signedTx = await tx.signAsync(user_bob);
                 await context.createBlock([signedTx]);
-                const events = await polkadotJs.query.system.events();
-                const fee = extractFeeAuthor(events, user_dave.address).amount.toBigInt();
 
-                // Gets the new pot deposit value
-                const new_pot = await polkadotJs.query.system.account(treasury_address);
-                const new_free_pot = new_pot.data.free.toBigInt();
-
-                // Division operation rounding
-                const rounding = fee % 5n > 0 ? 1n : 0n;
-
-                // Treasury pot should increase by 20% of the paid fee & tip
-                expect(new_free_pot).to.be.equal(initial_free_pot + fee / 5n + rounding);
+                expect((await polkadotJs.query.treasury.spendCount()).toNumber()).to.equal(0);
             },
         });
 
         it({
             id: "E02",
-            title: "Non root can not spend from treasury",
+            title: "Root can spend from treasury (Local)",
+            test: async function () {
+                expect((await polkadotJs.query.treasury.spendCount()).toNumber()).to.equal(0);
+                const balanceBefore = (await polkadotJs.query.system.account(user_dave.address)).data.free.toBigInt();
+
+                // Creates a proposal
+                // Value needs to be higher than the transaction fee paid by dave, but lower than the total treasury pot
+                const proposal_value = 1000000000n;
+                const tx = polkadotJs.tx.treasury.spendLocal(proposal_value, user_dave.address);
+                const signedTx = await polkadotJs.tx.sudo.sudo(tx).signAsync(sudo_alice);
+                await context.createBlock([signedTx]);
+
+                // Local spends dont upadte the spend count
+                expect((await polkadotJs.query.treasury.spendCount()).toNumber()).to.equal(0);
+
+                // After the payout is approved, we need to wait SpendPeriod for the payout to happen
+                const spendPeriod = polkadotJs.consts.treasury.spendPeriod;
+
+                // Now we just wait the spendPeriod, no need for payout calls in local spends
+                await jumpBlocks(context, spendPeriod.toNumber())
+
+                const balanceAfter = (await polkadotJs.query.system.account(user_dave.address)).data.free.toBigInt();
+                expect(balanceAfter).toBeGreaterThan(balanceBefore);
+            },
+        });
+
+        it({
+            id: "E03",
+            title: "Non root can not spend from treasury (Non-local)",
             test: async function () {
                 expect((await polkadotJs.query.treasury.spendCount()).toNumber()).to.equal(0);
 
@@ -71,8 +94,8 @@ describeSuite({
         });
 
         it({
-            id: "E03",
-            title: "Root can spend from treasury",
+            id: "E04",
+            title: "Root can spend from treasury (Non-local)",
             test: async function () {
                 expect((await polkadotJs.query.treasury.spendCount()).toNumber()).to.equal(0);
                 const balanceBefore = (await polkadotJs.query.system.account(user_dave.address)).data.free.toBigInt();
